y1apgm-django-multi-tenant-saas-api-with-row-level-security/requirements.txt
# Root requirements for all tests and evaluation
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.7.0
